Ensure we logout of bitwarden, even when an error has occurred
This fixes a bug in successive lambda runs due to disk persistence as calling logout runs a cleanup of the encrypted appdata on disk.
If an API call fails (due to timeout ect) the data remains in place. This results in login failing on the next invocation login as we are "already logged in".

I think it's better ensure a cleanup here rather than ignoring "already logged in" errors so that app data is as fresh as possible to work on.